move step2input and step2syscall to step subcommands
