+++ crates/mipsy_interactive/src/interactive/commands/mod.rs
@@ -13,8 +13,6 @@ mod print;
 mod reset;
 mod run;
 mod step;
-mod step2input;
-mod step2syscall;
 pub(crate) mod util;
 mod watchpoint;
 
@@ -32,8 +30,6 @@ pub(crate) use print::print_command;
 pub(crate) use reset::reset_command;
 pub(crate) use run::run_command;
 pub(crate) use step::step_command;
-pub(crate) use step2input::step2input_command;
-pub(crate) use step2syscall::step2syscall_command;
 pub(crate) use watchpoint::watchpoint_command;
 
 use super::{error::CommandResult, State};
