+++ crates/mipsy_interactive/src/interactive/commands/step.rs
@@ -7,27 +7,44 @@ use crate::prompt;
 use super::*;
 use super::Command;
 use colored::*;
+use mipsy_lib::Register;
 
 pub(crate) fn step_command() -> Command {
-    let subcommands = vec![command(
-        "back",
-        vec!["b"],
-        vec![],
-        vec!["times"],
-        vec![],
-        "",
-        |_, state, label, args| step_back(state, label, args),
-    )];
+    let subcommands = vec![
+        command(
+            "back",
+            vec!["b"],
+            vec![], vec!["times"], vec![],
+            "",
+            |_, state, label, args| step_back(state, label, args),
+        ),
+        command(
+            "syscall",
+            vec![""],
+            vec![], vec![], vec![],
+            "",
+            |_, state, label, args| step_syscall(state, label, args),
+        ),
+        command(
+            "input",
+            vec![""],
+            vec![], vec![], vec![],
+            "",
+            |_, state, label, args| step_input(state, label, args),
+        ),
+    ];
 
     // TODO:
+    //  - dedup step/back logic
     //  - back alias
+    //  - long help
     command(
         "step",
         vec!["s"],
         vec![],
-        vec!["times"],
+        vec!["times", "subcommand"],
         subcommands,
-        &format!("step forwards one (or {}) instruction", "[times]".magenta()),
+        &format!("step forwards or execute a subcommand"),
         |cmd, state, label, args| {
             if label == "__help__" && args.is_empty() {
                 return Ok(get_long_help());
@@ -149,6 +166,7 @@ fn step_back(state: &mut State, label: &str, args: &[String]) -> Result<String, 
 
     let mut backs = 0;
     let mut ran_out_of_history = false;
+    state.interrupted.store(false, Ordering::SeqCst);
     for _ in 0..times {
         let runtime = &mut state.runtime;
 
@@ -166,6 +184,10 @@ fn step_back(state: &mut State, label: &str, args: &[String]) -> Result<String, 
             state.exited = false;
         } else if backs == 0 {
             return Err(CommandError::CannotStepFurtherBack);
+        }
+
+        if state.interrupted.load(Ordering::SeqCst) {
+            break;
         }
     }
 
@@ -198,6 +220,93 @@ fn step_back(state: &mut State, label: &str, args: &[String]) -> Result<String, 
         );
     }
     println!();
+
+    Ok("".into())
+}
+
+fn step_syscall(state: &mut State, label: &str, _args: &[String]) -> Result<String, CommandError> {
+    if label == "__help__" {
+        return Ok(
+            format!(
+                "Steps forwards until your program's next syscall.\n\
+                 This will run in \"verbose\" mode, printing out each instruction that was\n\
+             \x20 executed, and verbosely printing any system calls that are executed.\n\
+                 To step backwards (i.e. back in time), use `{}`.",
+                "step back".bold(),
+            ),
+        )
+    }
+
+    if state.exited {
+        return Err(CommandError::ProgramExited);
+    }
+
+    state.interrupted.store(false, Ordering::SeqCst);
+    while !state.interrupted.load(Ordering::SeqCst) {
+        let binary  = state.binary.as_ref().ok_or(CommandError::MustLoadFile)?;
+        let runtime = &state.runtime;
+
+        let syscall = if let Ok(inst) = runtime.next_inst() {
+            util::print_inst(&state.iset, binary, inst, runtime.timeline().state().pc(), state.program.as_deref());
+
+            inst == 0xC
+        } else {
+            false
+        };
+
+        let step = state.step(true)?;
+
+        if step || syscall {
+            break;
+        }
+
+    }
+
+    Ok("".into())
+}
+
+fn step_input(state: &mut State, label: &str, _args: &[String]) -> Result<String, CommandError> {
+    if label == "__help__" {
+        return Ok(
+            format!(
+                "Steps forwards until your program asks for its next input, or finishes.\n\
+                 This will run in \"verbose\" mode, printing out each instruction that was\n\
+             \x20 executed, and verbosely printing any system calls that are executed.\n\
+                 To step backwards (i.e. back in time), use `{}`.",
+                "back".bold(),
+            ),
+        )
+    }
+
+    if state.exited {
+        return Err(CommandError::ProgramExited);
+    }
+
+    state.interrupted.store(false, Ordering::SeqCst);
+    while !state.interrupted.load(Ordering::SeqCst) {
+        let binary  = state.binary.as_ref().ok_or(CommandError::MustLoadFile)?;
+        let runtime = &state.runtime;
+
+        let input = if let Ok(inst) = runtime.next_inst() {
+            util::print_inst(&state.iset, binary, inst, runtime.timeline().state().pc(), state.program.as_deref());
+
+            if inst == 0xC {
+                let syscall = runtime.timeline().state().read_register(Register::V0.to_u32()).unwrap_or(-1);
+                matches!(syscall, 5 | 6 | 7 | 8 | 12)
+            } else {
+                false
+            }
+        } else {
+            false
+        };
+
+        let step = state.step(true)?;
+
+        if step || input {
+            break;
+        }
+
+    }
 
     Ok("".into())
 }
